requirements.txt
pytest~=8.3.5
requests~=2.32.3
Faker~=37.0.0
pydantic==2.10.6
python-dotenv~=1.0.1
SQLAlchemy~=2.0.39
allure-python-commons~=2.13.5